test(mfsdp): separate fully shard test concerns
